separated classes into different files
also added back the google fonts badge
