+++ src/demos/MagnifyDemo.js
